Add Holding model and a data importer
This gets data from the holding-list.csv into the Rails database.

Making a data directory and putting the holding-list.csv into it was a
rather arbitrary choice. In most circumstances in which you might want
something like this, the CSV would be outside of the Rails app. In this
case, I included it in the checked-in app so that it's just that much
easier to get this app in a working state. All you have to do is open a
console and run `DataImporter.run`.

Every field on Holding is required. It's easy to envision a data set in
which there is some missing data which we want to import anyway, but
with this data set, it's all there, and it's valuable to keep the
current good state of data "proven" until we have to back it out.

I did not yet add any additional validations besides presence
validations (such as numericality or length checks). Nor did I write any
tests yet.