Add notes about a holding's attributes to README
Did a little research on bonds and decided that belonged somewhere!

This also might belong as comments in the Holding model. That would
probably be a better place as the app (and things documented in the
README) begins to grow. But for now, the README isn't too unwieldy to be
a decent place.