Set up RSpec and write tests for index action
Gems added: RSpec, FactoryBot, Pry, Faker

Removed the test directory generated by Rails in favor of a spec directory.

Added a factory for sort of realistic random fake Holding data to use in
tests.

Created basic tests for the index action of the holdings API, showing
that it does seem to be outputting the right data.